fixed SourceDestinationMap to parse long, not int

+++ src/main/java/adventofcode/day05/SourceDestinationMap.java
@@ -8,9 +8,9 @@ public record SourceDestinationMap(long destinationStart, long sourceStart, long
     public static SourceDestinationMap createFromInputLine(String inputLine) {
         String[] parameters = inputLine.split(" ");
         return new SourceDestinationMap(
-                Integer.parseInt(parameters[0]),
-                Integer.parseInt(parameters[1]),
-                Integer.parseInt(parameters[2])
+                Long.parseLong(parameters[0]),
+                Long.parseLong(parameters[1]),
+                Long.parseLong(parameters[2])
         );
     }
 }
